Implement search throught context instead of props
